add test for relayChain
